# encode complex types as "%+v"